CanvasPalette 분리 및 이름 변경
